Added a naive .fasta sequence reader + a benchmark test that creates a large ngram called bioTest. Added CNE data and a simple svm classifier plus a testing svm classify program using normalized value similarity between human and rat CNE genomes.
