--- documentModel/__init__.py
@@ -1,2 +1,0 @@
-from representations import *
-from comparators import *
