--- documentModel/comparators/.cvsignore
@@ -1,1 +1,0 @@
-.svn
