--- documentModel/comparators/__init__.py
@@ -1,2 +1,0 @@
-from NGramGraphSimilarity import *
-from Operator import *
