--- documentModel/representations/__init__.py
@@ -1,3 +1,0 @@
-from DocumentNGramGaussNormGraph import *
-from DocumentNGramSymWinGraph import *
-from DocumentNGramGraph import *
