+++ source/__init__.py
@@ -1,0 +1,3 @@
+from bioRead import *
+from documentModel import *
+from classify import *
