+++ source/documentModel/__init__.py
@@ -1,0 +1,2 @@
+from representations import *
+from comparators import *
