+++ source/documentModel/comparators/__init__.py
@@ -1,0 +1,2 @@
+from NGramGraphSimilarity import *
+from Operator import *
