+++ source/documentModel/representations/__init__.py
@@ -1,0 +1,3 @@
+from DocumentNGramGaussNormGraph import *
+from DocumentNGramSymWinGraph import *
+from DocumentNGramGraph import *
